openrouter + consistent agent response with better prompting

+++ models.py
@@ -7,6 +7,6 @@ class UserProfile(BaseModel):
     location: Optional[str] = None
     property_type: Optional[str] = None
     bedrooms: Optional[int] = None
-    bathrooms: Optional[int] = None
+    bathrooms: Optional[float] = None
     must_haves: List[str] = []
     good_to_haves: List[str] = []
